Split out DomTracker
