+++ .editorconfig
@@ -1,0 +1,3 @@
+root = true
+indent_style = space
+indent_size = 4
